+++ kmeans.c
@@ -12,11 +12,11 @@
 
 #define MAX(x, y) ((x > y) ? x : y)
 
-static void update_means(double *U, int *indices, int k, int n, double *ret) {
+static void update_means(float *U, int *indices, int k, int n, float *ret) {
     ENTER_FUNC;
     NUM_ADDS(n*k);
     NUM_DIVS(k*k);
-    double *tmp_means = calloc(k * k, sizeof(double));
+    float *tmp_means = calloc(k * k, sizeof(float));
     int *sizes = calloc(k, sizeof(int));;
     for (int i = 0; i < n ; i++) { // iterate over each point
         for (int j = 0; j < k; j++) { // iterate over each indices
@@ -32,13 +32,13 @@ static void update_means(double *U, int *indices, int k, int n, double *ret) {
     EXIT_FUNC;
 }
 
-static int find_nearest_cluster_index(double *point, double *means, int k) {
+static int find_nearest_cluster_index(float *point, float *means, int k) {
     ENTER_FUNC;
     // use l2_norm
-    double gap = DBL_MAX;
+    float gap = DBL_MAX;
     int index = 0;
     for (int i = 0; i < k; i++) { // for every cluster check abs distance to point and take the minimal
-        double norm = l2_norm_squared(point, &means[i*k], k);
+        float norm = l2_norm_squared(point, &means[i*k], k);
         if(norm < gap) {
             gap = norm;
             index = i;
@@ -59,12 +59,12 @@ static int find_nearest_cluster_index(double *point, double *means, int k) {
  *   5. for each cluster j = 1..k
  *          - new centroid = mean of all points assigned to that cluster
  */
-void lloyd_kmeans(double *U, int n, int k, int max_iter, double stopping_error, struct cluster *ret) {
+void lloyd_kmeans(float *U, int n, int k, int max_iter, float stopping_error, struct cluster *ret) {
     ENTER_FUNC;
     // k is the number of columns in U matrix  U is a n by k matrix (here only!)
     int i = 0;
     // each row represents a cluster each column a dimension
-    double means[k*k];
+    float means[k*k];
     int indices[n];
     while (i < max_iter) {
         if (i < 1) {
@@ -145,8 +145,8 @@ int write_clustering_result(char *file, struct cluster *clusters, int num_cluste
  * 1) init DS
  * 2) init kpp
  */
-static void initialize(double *clusters_center, double *U, int *clusters_size, double *upper_bounds
-        , double *lower_bounds, int *cluster_assignments, int k, int n) {
+static void initialize(float *clusters_center, float *U, int *clusters_size, float *upper_bounds
+        , float *lower_bounds, int *cluster_assignments, int k, int n) {
     ENTER_FUNC;
     clusters_size[0] = n; // first contains all
     for (int i = 1; i < k; i++) {
@@ -168,14 +168,14 @@ static void initialize(double *clusters_center, double *U, int *clusters_size, d
  * 1) find the two closest centers,
  * 2) update the bounds if closest changed, the assignments and the cluster sizes
  */
-static void point_all_clusters(double *U, double *clusters_center, int *cluster_assignments
-        , double *upper_bounds, double *lower_bounds, int *clusters_size, int k, int i) {
+static void point_all_clusters(float *U, float *clusters_center, int *cluster_assignments
+        , float *upper_bounds, float *lower_bounds, int *clusters_size, int k, int i) {
     ENTER_FUNC;
     int closest_center_1 = 0;
-    double closest_center_1_dist = DBL_MAX;
-    double closest_center_2_dist = DBL_MAX;
+    float closest_center_1_dist = DBL_MAX;
+    float closest_center_2_dist = DBL_MAX;
     for (int j = 0; j < k; j++) {
-        double dist = l2_norm(U + i * k, clusters_center + j * k, k);
+        float dist = l2_norm(U + i * k, clusters_center + j * k, k);
         // Find distance between the point and the center.
         if (dist < closest_center_1_dist) {
             closest_center_2_dist = closest_center_1_dist;
@@ -204,12 +204,12 @@ static void point_all_clusters(double *U, double *clusters_center, int *cluster_
  * 2) reassign new centers
  * return maximal dist moved;
  */
-static double move_centers(double *new_clusters_centers, int *clusters_size, double *clusters_center
-        , double *centers_dist_moved, int k) {
+static float move_centers(float *new_clusters_centers, int *clusters_size, float *clusters_center
+        , float *centers_dist_moved, int k) {
     ENTER_FUNC;
-    double dist_moved = 0;
+    float dist_moved = 0;
     for (int j = 0; j < k; j++) {
-        double dist = 0;
+        float dist = 0;
         if (clusters_size[j] > 0) {
             for (int l = 0; l < k; l++) { // update
                 if (new_clusters_centers[j*k + l] == clusters_size[j]) {
@@ -234,8 +234,8 @@ static double move_centers(double *new_clusters_centers, int *clusters_size, dou
  * ALGO 5 - UPDATE BOUNDS ---------------------------------------------------------
  * 1) update the new bounds
  */
-static void update_bounds(double *upper_bounds, double *lower_bounds, double *centers_dist_moved
-        , int *cluster_assignments, double max_dist_moved, int n) {
+static void update_bounds(float *upper_bounds, float *lower_bounds, float *centers_dist_moved
+        , int *cluster_assignments, float max_dist_moved, int n) {
     ENTER_FUNC;
     for (int i = 0; i < n; i++) {
         NUM_ADDS(2);
@@ -250,26 +250,26 @@ static void update_bounds(double *upper_bounds, double *lower_bounds, double *ce
  * Implementation of the following algorithms as presented in the paper:
  *      https://epubs.siam.org/doi/pdf/10.1137/1.9781611972801.12
  */
-void hamerly_kmeans(double *U, int n, int k, int max_iter, double stopping_error, struct cluster *ret) {
+void hamerly_kmeans(float *U, int n, int k, int max_iter, float stopping_error, struct cluster *ret) {
     ENTER_FUNC;
     // initial centers
-    double clusters_center[k*k];
+    float clusters_center[k*k];
     // tmp for next iteration
-    double new_clusters_centers[k*k];
+    float new_clusters_centers[k*k];
     // cluster sizes
     int clusters_size[k];
     // n upper bounds (of closest center)
-    double upper_bounds[n];
+    float upper_bounds[n];
     // n lower bounds (of 2nd strict closest center)
-    double lower_bounds[n];
+    float lower_bounds[n];
     // stores cluster index for all points
     int cluster_assignments[n];
     // Algorithm 2: init + kpp -------------------
     initialize(clusters_center, U, clusters_size, upper_bounds, lower_bounds, cluster_assignments, k, n);
     // Distance to nearest other cluster for each cluster.
-    double dist_nearest_cluster[k];
+    float dist_nearest_cluster[k];
     // distance of centers moved between two iteration
-    double centers_dist_moved[k];
+    float centers_dist_moved[k];
     int iteration = 0;
     while (iteration < max_iter) {
         // Initialization after each iteration
@@ -278,10 +278,10 @@ void hamerly_kmeans(double *U, int n, int k, int max_iter, double stopping_error
         }
         // min distance between each two centers {update s} --------------------------
         for (int i = 0; i < k; i++) { // for each cluster
-            double min_dist = DBL_MAX;
+            float min_dist = DBL_MAX;
             for (int j = 0; j < k; j++) { // look at the distances to all cluster
                 if (i != j) { // is 0
-                    double dist = 0;
+                    float dist = 0;
                     for (int l = 0; l < k; l++) { // iterate over column = dimension
                         NUM_MULS(1);
                         NUM_ADDS(3);
@@ -301,7 +301,7 @@ void hamerly_kmeans(double *U, int n, int k, int max_iter, double stopping_error
         // ALGO 1: line 5
         for (int i = 0; i < n; i++) {
             // line 6: max_d = max(s(a(i))/2, l(i)) ???
-            double max_d = fmax(lower_bounds[i], dist_nearest_cluster[cluster_assignments[i]]);
+            float max_d = fmax(lower_bounds[i], dist_nearest_cluster[cluster_assignments[i]]);
             // ALGO 1: line7: {first bound test}
             if (upper_bounds[i] > max_d) {
                 upper_bounds[i] = l2_norm(U + i * k, clusters_center + cluster_assignments[i] * k, k);
@@ -321,12 +321,12 @@ void hamerly_kmeans(double *U, int n, int k, int max_iter, double stopping_error
             }
         }
         // ALGO 4 - MOVE-CENTERS: check for distance moved then move the centers ---------
-        double max_dist_moved = move_centers(new_clusters_centers, clusters_size
+        float max_dist_moved = move_centers(new_clusters_centers, clusters_size
                 , clusters_center, centers_dist_moved, k);
         // ALGO 5 - Update-bounds : for all U update upper and lower distance bounds ---------------
         update_bounds(upper_bounds, lower_bounds, centers_dist_moved, cluster_assignments, max_dist_moved, n);
         // transfer new state to current
-        memcpy(clusters_center, new_clusters_centers, k * k * sizeof(double));
+        memcpy(clusters_center, new_clusters_centers, k * k * sizeof(float));
         iteration++;
     }
     // write into convenient data-structure struct cluster
@@ -350,20 +350,20 @@ void hamerly_kmeans(double *U, int n, int k, int max_iter, double stopping_error
     EXIT_FUNC;
 }
 
-static void comp_distance_between_centers(double *means, int k, double *ret_dist_centers){
+static void comp_distance_between_centers(float *means, int k, float *ret_dist_centers){
     ENTER_FUNC;
     for(int i = 0; i < k; i++){
         for(int j = 0; j <= i; j++){
             //symmetric matrix, only store in the lower matrix
             //TODO  better way to store (might be not worth it, seems computing the index result in worse execution)
-            double dist = l2_norm(&means[i*k], &means[j*k], k);
+            float dist = l2_norm(&means[i*k], &means[j*k], k);
             ret_dist_centers[i*k+j] = dist;
         }
     }
     EXIT_FUNC;
 }
 
-static double get_dist_centers(int i, int j, double *dist_centers, int k){
+static float get_dist_centers(int i, int j, float *dist_centers, int k){
     if(i>j){
         return dist_centers[i*k+j];
     }else{
@@ -371,13 +371,13 @@ static double get_dist_centers(int i, int j, double *dist_centers, int k){
     }
 }
 
-static void comp_array_s(double *dist_centers, int k, double *ret_s){
+static void comp_array_s(float *dist_centers, int k, float *ret_s){
     for(int i =0; i < k; i++){
-        double max = DBL_MAX;
+        float max = DBL_MAX;
         for(int j = 0; j < k; j++){
             if(j!=i){
                 NUM_MULS(1);
-                double half_dist = 0.5*dist_centers[i*k+j];
+                float half_dist = 0.5*dist_centers[i*k+j];
                 if(half_dist < max){
                     max = half_dist;
                 }
@@ -388,10 +388,10 @@ static void comp_array_s(double *dist_centers, int k, double *ret_s){
 }
 
 
-static void init_elkan(double *U, int n, int k, double *means, double *lb, double *ub, int *indices){
+static void init_elkan(float *U, int n, int k, float *means, float *lb, float *ub, int *indices){
     ENTER_FUNC;
 
-    double dist_centers[k*k];
+    float dist_centers[k*k];
     //initialization
     for(int i = 0; i<k; i++){
         for(int j = 0; j<k; j++){
@@ -406,7 +406,7 @@ static void init_elkan(double *U, int n, int k, double *means, double *lb, doubl
         //then use lemma 1 in https://www.aaai.org/Papers/ICML/2003/ICML03-022.pdf
         //to avoid redundant computation
         int cur_cluster_assigned = 0;
-        double dist_p_c = l2_norm(&U[i * k], &means[0*k], k);
+        float dist_p_c = l2_norm(&U[i * k], &means[0*k], k);
         lb[0*k+0] = 0;
         ub[i] = dist_p_c;
 
@@ -415,7 +415,7 @@ static void init_elkan(double *U, int n, int k, double *means, double *lb, doubl
             NUM_MULS(1);
             if (0.5 * dist_centers[cur_cluster_assigned * k + j] < dist_p_c) {
                 //first compute distance between current point and cluster j's mean
-                double dist_p_c_j = l2_norm(&U[i * k], &means[j * k], k);
+                float dist_p_c_j = l2_norm(&U[i * k], &means[j * k], k);
                 lb[i * k + j] = dist_p_c_j;
                 if (dist_p_c_j < dist_p_c) {
                     //found a closer cluster
@@ -437,23 +437,23 @@ static void init_elkan(double *U, int n, int k, double *means, double *lb, doubl
  * ELKAN
  */
 
-void elkan_kmeans(double *U, int n, int k, int max_iter, double stopping_error, struct cluster *ret) {
+void elkan_kmeans(float *U, int n, int k, int max_iter, float stopping_error, struct cluster *ret) {
     // printf("entering kmeans\n");
     ENTER_FUNC;
     // k is the number of columns in U matrix  U is a n by k matrix (here only!)
 
     // each row represents a cluster each column a dimension
-    double *means = malloc(k*k*sizeof(double));
-    double *new_means = malloc(k*k*sizeof(double));
+    float *means = malloc(k*k*sizeof(float));
+    float *new_means = malloc(k*k*sizeof(float));
     // each row represents the lower bound of the same point x and different centers (total k centers)
-    double *lb = malloc(n*k*sizeof(double));
-    double *ub = malloc(n*sizeof(double));
+    float *lb = malloc(n*k*sizeof(float));
+    float *ub = malloc(n*sizeof(float));
     int *indices = malloc(n*sizeof(int));
 //    int *r = malloc(n*sizeof(int));
     int r;
-    double *dist_centers = malloc(k*k*sizeof(double));
-    double *s_dist_centers = malloc(k*sizeof(double));
-    double delta[k];
+    float *dist_centers = malloc(k*k*sizeof(float));
+    float *s_dist_centers = malloc(k*sizeof(float));
+    float delta[k];
 //    for(int i = 0; i < n ; i++){r[i] = 1;}
 
     init_kpp(&U[0], n, k, means);
@@ -474,7 +474,7 @@ void elkan_kmeans(double *U, int n, int k, int max_iter, double stopping_error, 
         //step 2 and 3
         for (int i = 0; i < n; i++){
             r = 1;
-            double dist_p_c = 0;
+            float dist_p_c = 0;
             if(ub[i] <= s_dist_centers[indices[i]]){
                 //do nothing?
             }else{
@@ -491,7 +491,7 @@ void elkan_kmeans(double *U, int n, int k, int max_iter, double stopping_error, 
                             }
                             NUM_MULS(1);
                             if((dist_p_c > lb[i*k+j]) || (dist_p_c > 0.5*get_dist_centers(indices[i], j, dist_centers, k))){
-                                double dist_p_j = l2_norm(&U[i * k], &means[j*k], k);
+                                float dist_p_j = l2_norm(&U[i * k], &means[j*k], k);
                                 if(dist_p_j < dist_p_c){
                                     indices[i] = j;
                                 }
@@ -512,9 +512,9 @@ void elkan_kmeans(double *U, int n, int k, int max_iter, double stopping_error, 
         for(int i = 0; i<n; i++){
             for(int j = 0; j<k;j++) {
 //                    todo double check this c, m(c)
-                double temp = lb[i * k + j];
-//                double norm_temp =l2_norm(&means[j * k], &new_means[j * k], k);
-//                /*double norm_temp =*/ l2_norm((double[]){1.0, 1.0, 987, 234, 123, 34},(double[]){123,123,123,123,123,123} , 6);
+                float temp = lb[i * k + j];
+//                float norm_temp =l2_norm(&means[j * k], &new_means[j * k], k);
+//                /*float norm_temp =*/ l2_norm((float[]){1.0, 1.0, 987, 234, 123, 34},(float[]){123,123,123,123,123,123} , 6);
                 lb[i * k + j] = MAX(0, temp - delta[j]);
 //                      lb[i*k+j] = 0;
             }
