+++ kmeans.h
@@ -2,14 +2,14 @@
 #define _KMEANS_H
 
 struct cluster {
-    double *mean; // center of the cluster
+    float *mean; // center of the cluster
     int size; // size of cluster points
     int *indices; // stores the indices of the points of U (stored row-ise)
 };
 
-void lloyd_kmeans(double *U, int n, int k, int max_iter, double stopping_error, struct cluster *ret);
-void hamerly_kmeans(double *U, int n, int k, int max_iter, double stopping_error, struct cluster *ret);
-void elkan_kmeans(double *U, int n, int k, int max_iter, double stopping_error, struct cluster *ret);
+void lloyd_kmeans(float *U, int n, int k, int max_iter, float stopping_error, struct cluster *ret);
+void hamerly_kmeans(float *U, int n, int k, int max_iter, float stopping_error, struct cluster *ret);
+void elkan_kmeans(float *U, int n, int k, int max_iter, float stopping_error, struct cluster *ret);
 void print_cluster_indices(struct cluster *clusters, int num_clusters);
 int write_clustering_result(char *file, struct cluster *clusters, int num_clusters);
